// packages/web/src/app/interfaces/events.ts
export interface IEvent {
    approved: number;
    comment: string;
    createdAt: string;
    end: string;
    googleId: string;
    googleMeetLink: string;
    id: number;
    isGoogleEvent: number;
    isRequest: number;
    primaryColor: string;
    requestType: string;
    secondaryColor: string;
    start: string;
    timeOffset: number;
    title: string;
    userId: number;
}